# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.Credentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  main.Event:
    properties:
      data:
        additionalProperties: true
        description: Rest of the fields should go here.
        type: object
      eventType:
        type: string
      service:
        type: string
      tags:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Logs Audit API documentation
  version: 1.0.0
paths:
  /auth:
    get:
      description: Brings a new token
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - BasicAuth: []
      summary: Brings a new token for the user
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: add new users
      parameters:
      - description: User credentials
        in: body
        name: Input
        schema:
          $ref: '#/definitions/main.Credentials'
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Add a new user to DB
      tags:
      - Auth
  /events:
    get:
      description: get documents
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: "2017-11-22"
        in: query
        name: timestamp
        type: string
      - description: the name of the service that sends the event
        in: query
        name: service
        type: string
      - description: the type of the event
        in: query
        name: eventType
        type: string
      - description: extra data to search in the event body
        in: query
        name: data
        type: string
      - description: metadata given from the service when stores the events
        in: query
        name: tags
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Brings documents according to the criteria
      tags:
      - Events
    post:
      description: get documents
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body (raw, json)
        in: body
        name: Input
        schema:
          $ref: '#/definitions/main.Event'
      responses:
        "201":
          description: Created
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
        "403":
          description: Forbidden
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Brings documents according to the criteria
      tags:
      - Events
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"